pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "faster-whisper-api"
version = "0.1.0"
description = "A lean FastAPI wrapper around faster-whisper (CT2) for local transcription on GPU or CPU."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Kwangsung Kim" }
]
keywords = ["speech", "stt", "asr", "whisper", "faster-whisper", "fastapi", "transcription"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Multimedia :: Sound/Audio :: Speech",
  "Framework :: FastAPI",
  "Operating System :: OS Independent",
]

dependencies = [
  "fastapi>=0.115.0,<1",
  "uvicorn[standard]>=0.30.0,<1",
  "faster-whisper>=1.0.0,<2",
  "pydub>=0.25.1,<1",
  "ffmpeg-python>=0.2.0,<1",
  "python-dotenv>=1.0.1,<2",
  "pydantic>=2.8,<3",
]

[project.optional-dependencies]
# 개발/테스트 편의
dev = [
  "ruff>=0.6.0",
  "black>=24.0.0",
  "isort>=5.13.0",
  "mypy>=1.10.0",
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "types-requests",
  "types-python-dateutil",
]

[project.urls]
Homepage = "https://github.com/your-org-or-user/Faster-Whisper-API"
Issues = "https://github.com/your-org-or-user/Faster-Whisper-API/issues"

# 패키지 디스커버리: app/ 이하 모듈을 포함
[tool.setuptools.packages.find]
include = ["app*"]
namespaces = false

# 패키지 데이터(환경파일 샘플 등) 포함
[tool.setuptools.package-data]
"app" = ["config/env/*"]

[tool.black]
line-length = 120
target-version = ["py310"]
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.ruff]
line-length = 120
target-version = "py310"
select = ["E", "F", "W", "I", "UP", "B"]
ignore = [
  "E203",  # black과 충돌할 가능성 있어서 무시
]
exclude = [
  ".git",
  ".mypy_cache",
  ".ruff_cache",
  ".venv",
  "build",
  "dist",
]

[tool.isort]
profile = "black"
line_length = 120
py_version = 310
known_first_party = ["app"]

[tool.pytest.ini_options]
addopts = [
  "--color=yes",
  "--durations=0",
  "--strict-markers",
  "--doctest-modules",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
log_cli = true
markers = [
  "slow: slow tests",
]
minversion = "6.0"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["app"]

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_lines = [
  "pragma: nocover",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = false
no_implicit_optional = true
check_untyped_defs = true
ignore_missing_imports = true
